+++ sql/hive/src/main/scala/org/apache/spark/sql/hive/HiveStrategies.scala
@@ -34,7 +34,6 @@ import org.apache.spark.sql.execution.command.{CreateTableCommand, DDLUtils}
 import org.apache.spark.sql.execution.datasources.{CreateTable, LogicalRelation}
 import org.apache.spark.sql.execution.datasources.parquet.{ParquetFileFormat, ParquetOptions}
 import org.apache.spark.sql.hive.execution._
-import org.apache.spark.sql.hive.orc.OrcFileFormat
 import org.apache.spark.sql.internal.{HiveSerDe, SQLConf}
 
 
@@ -205,7 +204,8 @@ case class RelationConversions(
         sessionCatalog.metastoreCatalog.convertToLogicalRelation(
           relation,
           options,
-          classOf[org.apache.spark.sql.hive.orc.OrcFileFormat], "orc")
+          classOf[org.apache.spark.sql.hive.orc.OrcFileFormat],
+          "orc")
       }
     }
   }
